Move some more functions around
